update reference test because of new metadata packets parsed. remove init in namespace packages.
