+++ src/ccsds/packets/europa_clipper/__init__.py
@@ -3,6 +3,7 @@ import pkgutil
 
 __all__ = []
 for loader, module_name, _ in pkgutil.walk_packages(__path__):
-    __all__.append(module_name)
-    _module = loader.find_module(module_name).load_module(module_name)
-    globals()[module_name] = _module
+    if "test" not in module_name:
+        __all__.append(module_name)
+        _module = loader.find_module(module_name).load_module(module_name)
+        globals()[module_name] = _module
